+++ packages/core-services/src/db/migrations/items/008.updateSwapTxnHistory/constants.ts
@@ -1,0 +1,5 @@
+export const providerData = {
+  id: 'changenow',
+  name: 'ChangeNow',
+  imageUrl: 'https://changenow.io/favicon.ico',
+};