+++ packages/cysync-core/src/context/swap/index.tsx
@@ -1,3 +1,6 @@
+import { getCoinSupport } from '@cypherock/coin-support';
+import { IcpSupport } from '@cypherock/coin-support-icp';
+import { coinFamiliesMap } from '@cypherock/coins';
 import { IAccount, IWallet, ISwapData } from '@cypherock/db-interfaces';
 import {
   ExchangeApp,
@@ -5,11 +8,11 @@ import {
 } from '@cypherock/sdk-app-exchange';
 import { ManagerApp } from '@cypherock/sdk-app-manager';
 import { hexToUint8Array } from '@cypherock/sdk-utils';
-import React, { useCallback, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useRef, useState } from 'react';
 
 import { ErrorActionMap, ErrorIconNameMap } from '~/constants/errors';
 import { DeviceTask, useDeviceTask, useMemoReturn } from '~/hooks';
-import { createExchange } from '~/services/swapService';
+import { createExchange, getProviderDetails } from '~/services/swapService';
 import { getDB } from '~/utils';
 import logger from '~/utils/logger';
 
@@ -25,6 +28,8 @@ export interface IProviderDetails {
   id: string;
   name: string;
   imageUrl: string;
+  complianceEmail?: string;
+  txnBaseURL?: string;
 }
 
 export interface IQuote {
@@ -71,6 +76,7 @@ export interface SwapContextInterface {
   setReceiveFlowValidTill: (d: number) => void;
   fillDetails: (params: IFillDetailsParams) => void;
   exchangeDetails?: IExchangeDetails;
+  providerDetails?: Record<string, IProviderDetails>;
   initiateExchange: (address: string) => Promise<void>;
   closeExchange: () => Promise<void>;
   resetIndex: number;
@@ -103,6 +109,8 @@ export const createCustomError = (heading: string, subtext?: string) => ({
 export const SwapProvider: React.FC<SwapProviderProps> = ({ children }) => {
   const [currentPage, setCurrentPage] = useState(SwapPage.DETAILS);
   const [globalError, setGlobalError] = useState<any>();
+  const [providerDetails, setProviderDetails] =
+    useState<Record<string, IProviderDetails>>();
 
   const toNextPage = () => {
     setCurrentPage(p => Math.min(SwapPage.STATUS, p + 1));
@@ -180,6 +188,20 @@ export const SwapProvider: React.FC<SwapProviderProps> = ({ children }) => {
     setQuote(selectedQuote);
   };
 
+  const fetchProviderDetails = useCallback(async () => {
+    try {
+      const res = await getProviderDetails();
+      const { providersDetails } = res.data;
+      if (res.status === 200) {
+        setProviderDetails(providersDetails);
+      } else {
+        throw Error('Invalid response from server');
+      }
+    } catch (error) {
+      logger.error('Could not fetch provider details', { error });
+    }
+  }, []);
+
   // give details to exchange app (init)
   // start receive flow
   // get details from exchange app (receive signature)
@@ -222,6 +244,22 @@ export const SwapProvider: React.FC<SwapProviderProps> = ({ children }) => {
     dontExecuteTask: true,
   });
 
+  const getRefundAddress = async (account: IAccount) => {
+    const coinSupport = getCoinSupport(account.familyId);
+    const refundAddress = await coinSupport.getAccountAddress({
+      account,
+    });
+    if (account.familyId === coinFamiliesMap.icp) {
+      const { accountId } = (
+        coinSupport as IcpSupport
+      ).getAddressDetailsFromPublicKey({
+        pubKey: refundAddress,
+      });
+      return accountId;
+    }
+    return refundAddress;
+  };
+
   const initiateExchange = async (address: string) => {
     try {
       const sig = await getSignature.run();
@@ -236,6 +274,9 @@ export const SwapProvider: React.FC<SwapProviderProps> = ({ children }) => {
       )
         throw new Error('Invalid prerequisite data');
 
+      // using fromAccount address as refund address
+      const refundAddress = await getRefundAddress(fromAccount);
+
       // give details to server
       const result = await createExchange({
         id: quote.id,
@@ -243,6 +284,7 @@ export const SwapProvider: React.FC<SwapProviderProps> = ({ children }) => {
         fromCurrency: fromAccount.assetId,
         toCurrency: toAccount.assetId,
         amount: quote.fromAmount,
+        refundAddress,
         receiverAddress: address,
         receiverAddressSignature: Buffer.from(sig.result.signature).toString(
           'hex',
@@ -317,6 +359,10 @@ export const SwapProvider: React.FC<SwapProviderProps> = ({ children }) => {
 
   const transactionId = useRef<string>();
 
+  useEffect(() => {
+    fetchProviderDetails();
+  }, []);
+
   const ctx = useMemoReturn({
     currentPage,
     toNextPage,
@@ -335,6 +381,7 @@ export const SwapProvider: React.FC<SwapProviderProps> = ({ children }) => {
     setReceiveFlowValidTill,
     fillDetails,
     exchangeDetails,
+    providerDetails,
     initiateExchange,
     closeExchange,
     markTransactionAsSwap,