+++ packages/cysync-core/src/dialogs/SwapDialog.tsx
@@ -11,6 +11,8 @@ import {
   GoldExternalLink,
   GraphSwitchSmallIcon,
   Image,
+  MessageBox,
+  parseLangTemplate,
   ScrollableContainer,
   SummaryBox,
   SummaryItemType,
@@ -19,6 +21,7 @@ import {
 } from '@cypherock/cysync-ui';
 import { SwapStatus } from '@cypherock/db-interfaces';
 import React, { FC, useMemo } from 'react';
+import { useSwap } from '~/context';
 
 import { SwapTransactionRowData, useSwapTransactions } from '~/hooks';
 import {
@@ -36,12 +39,15 @@ const textColorMap: Record<SwapStatus, any> = {
   success: 'success',
   failed: 'error',
   pending: 'warn',
+  hold: 'warn',
+  expired: 'error',
 };
 
 export const SwapDialog: FC<ISwapDialogProps> = ({ swap: swapSource }) => {
   const theme = useTheme();
   const dispatch = useAppDispatch();
-  const { displayedData } = useSwapTransactions();
+  const { providerDetails } = useSwap();
+  const { displayedData } = useSwapTransactions(providerDetails);
   const onClose = () => dispatch(closeDialog('swapDialog'));
 
   const lang = useAppSelector(selectLanguage);
@@ -235,10 +241,10 @@ export const SwapDialog: FC<ISwapDialogProps> = ({ swap: swapSource }) => {
           align="center"
           direction="column"
           height="full"
-          pr={0}
           pb={0}
+          px={0}
         >
-          <Container align="center" justify="center" width="full">
+          <Container align="center" justify="center" width="full" px={5}>
             <GraphSwitchSmallIcon
               width={36}
               height={36}
@@ -251,6 +257,7 @@ export const SwapDialog: FC<ISwapDialogProps> = ({ swap: swapSource }) => {
             align="center"
             width="full"
             gap={4}
+            px={5}
           >
             <Typography variant="h5">
               {swap.sentDisplayAmount} → {swap.receivedDisplayAmount}
@@ -259,15 +266,35 @@ export const SwapDialog: FC<ISwapDialogProps> = ({ swap: swapSource }) => {
               {swap.dateHeader} {swap.time}
             </Typography>
           </Container>
+          {swap.swapStatus === SwapStatus.Hold && swap.providerMail && (
+            <Container
+              display="flex"
+              direction="column"
+              align="center"
+              width="full"
+              gap={4}
+              px={5}
+            >
+              <MessageBox
+                text={parseLangTemplate(
+                  lang.strings.swap.swapStatus.messageBox.hold,
+                  {
+                    providerName: swap.providerName,
+                    providerEmail: `[**${swap.providerMail}**](mailto:${swap.providerMail})`,
+                  },
+                )}
+                type="warning"
+              />
+            </Container>
+          )}
           <ScrollableContainer $maxHeight="calc(100vh - 400px)">
             <Container
               display="flex"
               direction="column"
               width="full"
-              pt={5}
-              pr={5}
               pb={3}
               gap={12}
+              px={5}
             >
               <SummaryBox items={summaryItems} />
             </Container>